pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secret-keeping"
version = "0.1.0"
description = "Secret keeping project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jinja2",
    "tqdm",
    "docent-python==0.1.27a0",
    "matplotlib-inline",
    "cloudpickle",
    "torchtyping",
    "torch==2.7.0",
    "openai>=2.17.0",
    "claude-agent-sdk",
    "fastmcp",
    "fastapi",
    "uvicorn",
    "ipykernel",
    "nest_asyncio",
    "petri @ git+https://github.com/safety-research/petri@abhay/auditing-agents",
]

[project.optional-dependencies]
dev = []

[tool.setuptools]
packages = ["src"]
